Adding lombok, renaming test packages, removed unused jobs folder
