+++ gradle.properties
@@ -15,6 +15,7 @@ awsSDKVersion=1.11.762
 bucket4jVersion=4.4.1
 commonsCLIVersion=1.4
 flinkScalaVersion=2.12
+gradleLombokPluginVersion=4.0.0
 guavaVersion=28.2-jre
 hadoopVersion=3.1.0
 jacksonDatabindVersion=2.9.1
@@ -22,6 +23,7 @@ junitVersion=4.13
 junitJupiterVersion = 5.7.1
 logbackVersion=1.2.3
 logstashLogbackEncoderVersion=4.11
+lombokVersion=1.18.12
 nettyBoringSSLVersion=2.0.17.Final
 nettyVersion=4.1.30.Final
 parquetVersion=1.10.0
