+++ lombok.config
@@ -1,0 +1,5 @@
+config.stopBubbling = true
+lombok.extern.findbugs.addSuppressFBWarnings = true
+#see https://github.com/rzwitserloot/lombok/issues/1612
+lombok.anyConstructor.addConstructorProperties=true
+lombok.addLombokGeneratedAnnotation = true
